backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
fair-esm==2.0.0
torch>=2.0.0
biopython==1.82
requests==2.31.0
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.4
pydantic==2.5.2
